+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.0.5] / 2022-01-24
+- Add SetImage / SetLongDescription
 - Add UpdateRibbonDescription
 - Add RibbonDescription